% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfastRaster.R
\name{bfastRaster}
\alias{bfastRaster}
\title{Run BFAST on an entire raster grid}
\usage{
bfastRaster(x, dates = NULL, start = NULL, end = NULL, interpolate,
  aggregate = "biweekly", f = 1, fast_bfast = TRUE, cores,
  h = 0.15, season = c("dummy", "harmonic", "none"), max.iter = NULL,
  breaks = NULL, hpc = "none", level = 0.05, type = "OLS-MOSUM")
}
\arguments{
\item{x}{Character vector. Full path names to images or to parent folders containing images.}

\item{start}{(Optional) Numeric or character. The start of the period of interest in the format \code{c(year, julian day)}.}

\item{end}{(Optional) Numeric or character. The end of the period of interest in the format \code{c(year, julian day)}.}

\item{interpolate}{Character. Either "linear" or "periodic" (see details).}

\item{aggregate}{Character. Either "weekly" , "biweekly" (also "fortnightly"), "monthly", "quarterly", or "yearly" (also "annually").}

\item{f}{Numeric. Factor by which to rescale values before running \code{bfast}. Defaults to 1 (no rescaling)}

\item{fast_bfast}{Logical. Enabling performance optimisations. This requires package \pkg{bfast} to be installed from a specific fork on \href{https://github.com/bfast2/bfast}{github}. If \code{FALSE}, the default implementation of \pkg{bfast} is used.}

\item{h}{minimal segment size between potentially detected breaks in the
trend model given as fraction relative to the sample size (i.e. the minimal
number of observations in each segment divided by the total length of the
timeseries.}

\item{season}{the seasonal model used to fit the seasonal component and
detect seasonal breaks (i.e. significant phenological change).  There are
three options: "dummy", "harmonic", or "none" where "dummy" is the model
proposed in the first Remote Sensing of Environment paper and "harmonic" is
the model used in the second Remote Sensing of Environment paper (See paper
for more details) and where "none" indicates that no seasonal model will be
fitted (i.e. St = 0 ). If there is no seasonal cycle (e.g. frequency of the
time series is 1) "none" can be selected to avoid fitting a seasonal model.}

\item{max.iter}{maximum amount of iterations allowed for estimation of
breakpoints in seasonal and trend component.}

\item{breaks}{integer specifying the maximal number of breaks to be
calculated. By default the maximal number allowed by h is used.}

\item{hpc}{A character specifying the high performance computing support.
Default is "none", can be set to "foreach". Install the "foreach" package
for hpc support.}

\item{level}{numeric; threshold value for the \link[strucchange]{sctest.efp}
test; if a length 2 vector is passed, the first value is used for the trend,
the second for the seasonality}

\item{type}{character, indicating the type argument to
\link[strucchange]{efp}}
}
\value{
A rasterBrick with layers depending on what has been supplied to \code{returnLayers}. 
By default, 3 layers are returned: 
(1) breakpoint: timing of breakpoints detected for each pixel; 
(2) magnitude: the median of the residuals within the monitoring period; 
(3) error: a value of 1 for pixels where an error was encountered by the algorithm (see \code{\link{try}}), and NA where the method was successfully run. See \code{\link{bfastmonitor}} for more information on the other possible layers.
}
\description{
Apply \code{\link[bfast]{bfast}} on a entire raster (Landsat time series).
}
\details{
\strong{Overview}

\code{bfastRaster} is theoretically designed to work on any generic raster time series, as long 
as a \code{dates} vector is provided. In the absence of a \code{dates} vector, \code{names(x)} 
should correspond exactly to respective Landsat scene ID's. In this case, 
\code{\link{Lmetadata}} is used to extract a dates vector, and subset by sensor if desired.

\strong{Arguments}

Arguments \code{start} and \code{end} adjust the period analysed. The time series is trimmed, 
butonly after interpolation (see argument \code{interpolate}).

Argument \code{interpolate} defines the method used to interpolate missing data. It
can be either 'linear' or 'periodic'. These methods use \code{\link[zoo]{na.approx}} and 
\code{\link[forecast]{na.interp}}, respectively (inspired in 
\href{https://philippgaertner.github.io/2018/04/bfast-preparation/}{this blog}).

Argument \code{aggregate} defines how the interpolated data (see argument \code{interpolate}),
which are daily time series, are aggregated to a different time frequency: weekly" , "biweekly"
(also "fortnightly"), "monthly", "quarterly", or "yearly" (also "annually"). The funtions used
for aggregate were copied from 
\href{https://philippgaertner.github.io/2018/04/bfast-preparation/}{this same blog}).

Arguments \code{h}, \code{season}, \code{max.iter}, \code{breaks}, \code{hpc}, \code{level}, and
\code{type} are inherited from \code{\link[bfast]{bfast}}. Note, however, that although argument
\code{h} is expected to be <= 0.5 in \code{\link[bfast]{bfast}}, values >=1 are accepted here. 
When so, \code{h} is interpreted as the minimal segment size in \strong{absolute} value. For 
example, in a monthly analysis, \code{h=12} means that the minimal segment size is 12 months.
Then, \code{h} is divided by the total length of the time series so to produce a relative 
fraction as expected by \code{\link[bfast]{bfast}}.
}
\examples{

\dontrun{
data(fire)
x<-fire
dates=NULL
#cell=8761217 # not used in this function
#start=c(2015,1)
#end=c(2018,365)
interpolate="periodic"
aggregate="weekly"
f=1
sensor=NULL
fast_bfast=TRUE
cores=2
#min.thresh=NULL # used in bfastPixel only
h=0.30952380952381
season = c('harmonic')
max.iter=10
breaks = NULL
hpc = 'none'
level=1
type= 'OLS-MOSUM'


data(fire)
system.time(
  out<-bfastRaster(fire, interpolate="linear", h=0.30952380952381, season = c('harmonic'), level=1, max.iter=5, fast_bfast=TRUE, cores=2)
)
plot(out)

bfastRaster(x, interpolate="periodic", cell=8761217, plot=TRUE, h=52, level=1, max.iter=5)
bfastRaster(x, interpolate="periodic", aggregate="weekly", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
bfastRaster(x, interpolate="periodic", aggregate="monthly", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
bfastRaster(x, interpolate="periodic", aggregate="quarterly", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
bfastRaster(x, start=c(2015,1), end=c(2018,365), interpolate="periodic", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
}

}
\seealso{
\code{\link{bfastPixel}}
}
\author{
Hugo Costa, Loic Dutrieux, Ben DeVries
}
