% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-aggregate.R
\name{.DGTaggregate}
\alias{.DGTaggregate}
\title{Internal functions}
\usage{
.aggregate.time.series(x, aggregate)
}
\arguments{
\item{x}{Regular time series.}

\item{aggregate}{Character. See \code{\link{bfastPixel}}.}
}
\value{
Aggregated time series
}
\description{
Functions to aggregate daily time series (copied from
\href{https://philippgaertner.github.io/2018/04/bfast-preparation/}{here})
}
\author{
Hugo Costa
}
\keyword{internal}
