% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLandsat.R
\name{.isLandsat}
\alias{.isLandsat}
\title{Check if a String is a Landsat file}
\usage{
.isLandsat(x)
}
\arguments{
\item{x}{Character}
}
\value{
\code{TRUE} if all elements of \code{x} match Landsat scene ID criteria, or \code{FALSE} otherwise.
}
\description{
Parses a vector of characters to check if they conform to Landsat naming convention
}
\examples{
sourcefile<-c("LC08_L1TP_204032_20190925_20190925_01_RT_sr_band1", "LC08_L1TP_204032_20160425_20170326_01_T1_sr_band1.tif")
DGTbfast:::.isLandsat(sourcefile)

sourcefile<-c("LC08_L1TP_204032_20190925_20190925_01_RT_sr_band1", "LC08_L1TP_204032")
DGTbfast:::.isLandsat(sourcefile)

}
\author{
Hugo Costa
}
