% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lmetadata.R
\name{Lmetadata}
\alias{Lmetadata}
\title{Retrieve Landsat info from filenames}
\usage{
Lmetadata(x, ...)
}
\arguments{
\item{x}{Character vector or Raster* objec. If Raster* object, its \code{\link[raster]{names}} will be used.}

\item{...}{Additional arguments to pass to \code{\link[utils]{write.csv}}.}
}
\value{
\code{data.frame}
}
\description{
Retrieves information on sensor, pathrow, date, collection and tier from the 
\href{https://www.usgs.gov/faqs/what-naming-convention-landsat-collections-level-1-scenes?qt-news_science_products=0#qt-news_science_products}{Landsat Collections Level-1 scenes naming convention}.
}
\examples{
\dontrun{
sourcefile<-c("LC08_L1TP_204032_20190925_20190925_01_RT_sr_band1", "LC08_L1TP_204032_20160425_20170326_01_T1_sr_band1.tif")
Lmetadata(sourcefile)
}

}
\author{
Hugo Costa
}
