% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfastPixel.R
\name{bfastPixel}
\alias{bfastPixel}
\title{Apply BFAST on a single pixel}
\usage{
bfastPixel(x, dates = NULL, start = NULL, end = NULL, interpolate,
  aggregate = "biweekly", cell = NULL, interactive = FALSE, f = 1,
  plot = TRUE, h = 0.15, season = c("dummy", "harmonic", "none"),
  max.iter = NULL, breaks = NULL, hpc = "none", level = 0.05,
  type = "OLS-MOSUM")
}
\arguments{
\item{x}{Character vector. Full path names to images or to parent folders containing images.}

\item{start}{(Optional) Numeric or character. The start of the period of interest in the format \code{c(year, julian day)}.}

\item{end}{(Optional) Numeric or character. The end of the period of interest in the format \code{c(year, julian day)}.}

\item{interpolate}{Character. Either "linear" or "periodic" (see details).}

\item{aggregate}{Character. Either "weekly" , "biweekly" (also "fortnightly"), "monthly", "quarterly", or "yearly" (also "annually").}

\item{cell}{Numeric. Either (i) a numeric of length 1 indicating the cell ID to be observed, or (ii) a numeric of length 2 representing the (x,y) coordinate of the raster cell to be observed. If \code{NULL}, argument \code{interactive} must be set to \code{TRUE}.}

\item{interactive}{Logical. Select cell by clicking on an already plotted map? (see \code{\link[raster]{click}}). If \code{FALSE}, \code{cell} cannot be \code{NULL}.}

\item{f}{Numeric. Factor by which to rescale values before running \code{bfast}. Defaults to 1 (no rescaling)}

\item{plot}{Logical. Plot the result? Defaults to \code{FALSE}.}

\item{h}{minimal segment size between potentially detected breaks in the
trend model given as fraction relative to the sample size (i.e. the minimal
number of observations in each segment divided by the total length of the
timeseries.}

\item{season}{the seasonal model used to fit the seasonal component and
detect seasonal breaks (i.e. significant phenological change).  There are
three options: "dummy", "harmonic", or "none" where "dummy" is the model
proposed in the first Remote Sensing of Environment paper and "harmonic" is
the model used in the second Remote Sensing of Environment paper (See paper
for more details) and where "none" indicates that no seasonal model will be
fitted (i.e. St = 0 ). If there is no seasonal cycle (e.g. frequency of the
time series is 1) "none" can be selected to avoid fitting a seasonal model.}

\item{max.iter}{maximum amount of iterations allowed for estimation of
breakpoints in seasonal and trend component.}

\item{breaks}{integer specifying the maximal number of breaks to be
calculated. By default the maximal number allowed by h is used.}

\item{hpc}{A character specifying the high performance computing support.
Default is "none", can be set to "foreach". Install the "foreach" package
for hpc support.}

\item{level}{numeric; threshold value for the \link[strucchange]{sctest.efp}
test; if a length 2 vector is passed, the first value is used for the trend,
the second for the seasonality}

\item{type}{character, indicating the type argument to
\link[strucchange]{efp}}
}
\value{
A list with the following components:
\enumerate{
  \item $bfast - an object of class 'bfast' (see \code{\link[bfast]{bfast}})
  \item $cell - the cell index (an integer of length 1). This can be used to run 
  \code{bfastPixel} again on the same pixel (with different parameters) without having to click
  on a plot again to find the same pixel (in that case, be sure to set 
  \code{interactive=FALSE} for subsequent trials!).
}
}
\description{
Apply \code{\link[bfast]{bfast}} on a single pixel of known cell index or coordinates, or 
interactively by clicking on a plot.
}
\details{
\strong{Overview}

\code{bfastPixel} is theoretically designed to work on any generic raster time series, as long 
as a \code{dates} vector is provided. In the absence of a \code{dates} vector, \code{x} or 
\code{names(x)} should correspond to respective Landsat filenames (see the
\href{https://www.usgs.gov/faqs/what-naming-convention-landsat-collections-level-1-scenes?qt-news_science_products=0#qt-news_science_products}{Landsat Collections Level-1 scenes naming convention}).
In this case, \code{\link{Lmetadata}} is used to extract a dates vector.

\strong{Arguments}

Arguments \code{start} and \code{end} adjust the period analysed. The time series is trimmed, 
but only after interpolation (see argument \code{interpolate}).

Argument \code{interpolate} defines the method used to interpolate missing data. It
can be either 'linear' or 'periodic'. These methods use \code{\link[zoo]{na.approx}} and 
\code{\link[forecast]{na.interp}}, respectively (inspired in 
\href{https://philippgaertner.github.io/2018/04/bfast-preparation/}{this blog}).

Argument \code{aggregate} defines how the interpolated data (see argument \code{interpolate}),
which are daily time series, are aggregated to a different time frequency: "weekly" , "biweekly"
(also "fortnightly"), "monthly", "quarterly", or "yearly" (also "annually"). The funtions used
for aggregate were copied from 
\href{https://philippgaertner.github.io/2018/04/bfast-preparation/}{this same blog}).

Arguments \code{h}, \code{season}, \code{max.iter}, \code{breaks}, \code{hpc}, \code{level}, and
\code{type} are inherited from \code{\link[bfast]{bfast}}. Note, however, that although argument
\code{h} is expected to be <= 0.5 in \code{\link[bfast]{bfast}}, values >=1 are accepted here. 
When so, \code{h} is interpreted as the minimal segment size in \strong{absolute} value. For 
example, in a monthly analysis, \code{h=12} means that the minimal segment size is 12 months.
Then, \code{h} is divided by the total length of the time series so to produce a relative 
fraction as expected by \code{\link[bfast]{bfast}}. This relative fraction is printed out as a 
warning.
}
\examples{
\dontrun{
x<-list.files("//DGT-692/Dgt-692- externo/Landsat8/Preprocessed/204032", "_preprocessed_NDVI.tif$",  full.names = TRUE, recursive = TRUE)

bfastPixel(x, interpolate="linear", cell=8761217, plot=TRUE, h=52, level=1, max.iter=5)
bfastPixel(x, interpolate="periodic", cell=8761217, plot=TRUE, h=52, level=1, max.iter=5)
bfastPixel(x, interpolate="periodic", aggregate="weekly", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
bfastPixel(x, interpolate="periodic", aggregate="monthly", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
bfastPixel(x, interpolate="periodic", aggregate="quarterly", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
bfastPixel(x, start=c(2015,1), end=c(2018,365), interpolate="periodic", cell=8761217, plot=TRUE, h=0.30952380952381, level=1, max.iter=5)
}

}
\seealso{
\code{\link{bfastRaster}}
}
\author{
Hugo Costa and Ben DeVries
}
